feat: Scaffold initial microservice architecture with project files, database migrations, and core service components.

+++ InventoryService.Application/Commands/CreateProductCommand.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace InventoryService.Application.Commands;
 
-public record CreateProductCommand(CreateProductDto Dto) : IRequest<long>;
+public record CreateProductCommand(CreateProductDto Dto) : IRequest<int>;
